pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vision-agent"
version = "0.1.0"
description = "AI that can see and interact with any UI using vision-language models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Edwin Isac", email = "edwinisac007@gmail.com" }
]
keywords = ["ai", "computer-vision", "automation", "ui", "gpt-4v", "vision", "rpa"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
desktop = ["pyautogui>=0.9.53"]
browser = ["playwright>=1.40.0"]
all = ["pyautogui>=0.9.53", "playwright>=1.40.0"]

[project.scripts]
vision-agent = "vision_agent.cli:main"

[project.urls]
Homepage = "https://github.com/edwiniac/vision-agent"
Repository = "https://github.com/edwiniac/vision-agent"
Issues = "https://github.com/edwiniac/vision-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["vision_agent"]
